Moved all TF related functions to tfs.cpp
